+++ index.html
@@ -228,6 +228,6 @@
   </template>
 
   <script src="js/functions.js" defer></script>
-  <script src="js/main.js" defer></script>
+  <script src="js/main.js" type="module" defer></script>
 </body>
 </html>
